fix: redux state must be unmutable
