+++ src/components/Comment/Emotionbox.js
@@ -1,12 +1,45 @@
 import styles from 'styles/components/comment/emotion.module.scss';
-import PropTypes from 'prop-types';
 import { toggleMyEmojis } from 'reducers/emojis';
-import { useDispatch } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
 
-export const EmotionBox = (props) => {
-  const { emotionList } = props;
+export const EmotionBox = () => {
   const dispatch = useDispatch();
+  const emojis = useSelector((state) => state.emojis.emojis);
+  const myEmojis = useSelector((state) => state.emojis.myEmojis);
 
+  const emotionList = [
+    {
+      count: emojis.fire,
+      isOn: myEmojis.fire,
+      text: 'Superb',
+      img: '/images/components/emotion/superb.svg',
+    },
+    {
+      count: emojis.like,
+      isOn: myEmojis.like,
+      text: 'Love',
+      img: '/images/components/emotion/love.svg',
+    },
+    {
+      count: emojis.wow,
+      isOn: myEmojis.wow,
+      text: 'Wow',
+      img: '/images/components/emotion/wow.svg',
+    },
+    {
+      count: emojis.sad,
+      isOn: myEmojis.sad,
+      text: 'Sad',
+      img: '/images/components/emotion/sad.svg',
+    },
+    {
+      count: emojis.fun,
+      isOn: myEmojis.fun,
+      text: 'Laugh',
+      img: '/images/components/emotion/laugh.svg',
+    },
+  ];
+  console.log(emotionList);
   const onClick = (type) => {
     let typ = '';
     if (type === 'Superb') {
@@ -45,7 +78,3 @@ export const EmotionBox = (props) => {
 };
 
 export default EmotionBox;
-
-EmotionBox.propTypes = {
-  emotionList: PropTypes.array,
-};
