+++ src/reducers/emojis.js
@@ -27,6 +27,7 @@ const typToType = {
 
 export const SET_EMOJIS = 'SET_EMOJIS';
 export const SET_MY_EMOJIS = 'SET_MY_EMOJIS';
+export const SET_ALL_EMOJIS = 'SET_ALL_EMOJIS';
 export const TOGGLE_EMOJI = 'TOGGLE_EMOJI';
 
 // actions
@@ -46,9 +47,14 @@ export const toggleEmoji = (data) => ({
   data,
 });
 
+export const setAllEmojis = (data) => ({
+  type: SET_ALL_EMOJIS,
+  data,
+});
+
 // API functions
 export const getMyEmojis = async (dispatch, id) => {
-  return EmojiAPI.getMyEmojis(id).then((res) => {
+  EmojiAPI.getMyEmojis(id).then((res) => {
     const myEmoji = { id };
     res.data.results.forEach((item) => {
       myEmoji[typToType[item.typ]] = true;
@@ -64,10 +70,20 @@ export const toggleMyEmojis = async (dispatch, getState, type) => {
   };
   const { emojis, myEmojis } = getState().emojis;
 
-  return EmojiAPI.toggleEmoji(myEmojis.id, data).then(() => {
-    emojis[typToType[type]] = emojis[typToType[type]] + myEmojis[typToType[type]] ? -1 : 1;
-    myEmojis[typToType[type]] = !myEmojis[typToType[type]];
-    dispatch(setMyEmojis(myEmojis));
+  EmojiAPI.toggleEmoji(myEmojis.id, data).then(() => {
+    const add = myEmojis[typToType[type]] ? -1 : 1;
+    dispatch(
+      setAllEmojis({
+        emojis: {
+          ...emojis,
+          [typToType[type]]: emojis[typToType[type]] + add,
+        },
+        myEmojis: {
+          ...myEmojis,
+          [typToType[type]]: !myEmojis[typToType[type]],
+        },
+      })
+    );
   });
 };
 
@@ -89,31 +105,10 @@ export const applySetMyEmojis = (state, action) => {
   };
 };
 
-export const applyCreateEmoji = (state, action) => {
+export const applySetAllEmojis = (state, action) => {
   return {
     ...state,
-    emojis: {
-      ...state.emojis,
-      [action.data.type]: state.emojis[action.data.type] + 1,
-    },
-    myEmojis: {
-      ...state.myEmojis,
-      [action.data.type]: action.data,
-    },
-  };
-};
-
-export const applyDeleteEmoji = (state, action) => {
-  return {
-    ...state,
-    emojis: {
-      ...state.emojis,
-      [action.data.type]: state.emojis[action.data.type] - 1,
-    },
-    myEmojis: {
-      ...state.myEmojis,
-      [action.data.type]: null,
-    },
+    ...action.data,
   };
 };
 
@@ -124,6 +119,8 @@ const reducer = (state = initialState, action) => {
       return applySetEmojis(state, action);
     case SET_MY_EMOJIS:
       return applySetMyEmojis(state, action);
+    case SET_ALL_EMOJIS:
+      return applySetAllEmojis(state, action);
     default:
       return state;
   }
